layout cell

+++ WClient/SYBWeiboAPIClient.h
@@ -25,8 +25,8 @@ typedef void(^PBEmptyBlock)();
 
 + (instancetype)sharedClient;
 
-- (void)getAllFriendsWeibo:(long)since_id
-                    max_id:(long)max_id
+- (void)getAllFriendsWeibo:(long long)since_id
+                    max_id:(long long)max_id
                       count:(int)count
                   base_app:(int)app_typ
                    feature:(int)feature
